Disabled edittext other than first row and added country flags

+++ app/src/main/java/com/revolut/currencyrate/model/Rate.kt
@@ -104,4 +104,4 @@ data class Rate(
 
 )
 
-data class RateItem(var rateKey: String, var rateValue: Double)
+data class RateItem(var rateKey: String, var rateValue: Float)
