+++ app/src/main/java/com/revolut/currencyrate/model/Rates.kt
@@ -15,6 +15,6 @@ data class Rates(
 
     @SerializedName("rates")
     @Expose
-    val rateList: Map<String, Double>? = null
+    val rateList: Map<String, Float>? = null
 
 )
